Bag of Tricks Enhancement Patch v1.3
This patch changes the behavior of the Bag of Tricks to something a
little more interesting than a wand of create monster in bag form.

Save/bones file compatibility is not affected.

As always, thanks to all the folks of rec.games.roguelike.nethack who
gave ideas and suggestions.

When you apply a bag of tricks a charge is used and one of the
following effects occurs:

d20     Effect
1-2     An item falls out of the bag.
        The item will be randomly generated if the bag has not
        been recharged and is not cursed, otherwise it will be an
        iron chain.
3       The bag wriggles away from you.
4       You are paralyzed for 1d4 turns. ("Something tries to pull
        you into the bag!")
5       Stinking cloud, centered at the bag.
6       Hallucination for 1d35+10 turns. ("The bag exhales a puff
        of spores.")
7       Scare monster effect. ("The bag yells "Boo!")
        If the bag is cursed or you fail a save vs wisdom, you are
        paralyzed for 1d3 turns. If the bag is blessed or you make
        your save, there is no effect on you.
8       The bag bites you for 1d10 points of damage.
9       The bag sprouts a tongue and flicks your weapon to the
        floor.
10-20   Monsters are created.

When the bag is depleted of charges, it can be used to store
items.

Looting the bag will get you bitten (as in Vanilla) unless it
has no charges, in which case the behavior is the same as
looting a normal container.

If you recharge the bag while items are in it, they get devoured.
("You hear loud crunching sounds from the bag.")

Side effects of existing code:
- The same things that destroy a bag of holding will also destroy a
  bag of tricks.
- A cursed bag of tricks will devour items, just like a cursed bag
  of holding.

In order to partially compensate for lower monster generation, a bag
of tricks will be generated with d30 charges instead of d20.

http://zindorsky.kundor.org/patches/tricks-desc.html

Signed-off-by: Nephi <zindorsky@gmail.com>
Signed-off-by: Steven Noonan <steven@uplinklabs.net>