disable gzip compression
The gzip path differs on Linux/Mac, which makes it a bit difficult
to pull off correctly.

On Mac OS X, having the wrong path caused a terrible bug. If you
closed the terminal window with NetHack running in it, then NetHack
would stay running in the background, on a --MORE-- prompt, taking
100% CPU.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>