Sortloot patch (28.6.2006)
This patch optionally causes item lists be sorted alphabetically by
description when you loot or apply containers or pick up something,
and with the "full" setting also in inventory and drop listings.
The items are still grouped in classes if the sortpack option is
on, but the order of items of the same class is different.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>