Scatter contents of bag of holding explosion instead of destroying it (from SporkHack rev. 525,526)
Quote SporkHack changelog:

Be nice to players for a change; just throw their stuff all over the place if
they blow up their bag of holding. ...Of course, it might hit hard enough to
break it. And it might fall in lava. Or water. And it can't be any fun to try
to pick all that stuff back up. :D


git-svn-id: https://unnethack.svn.sourceforge.net/svnroot/unnethack/trunk@943 f1e491ac-5ded-4967-80bf-9f4855be876a